Extracts the type reflection logic into a dedicated class
